+++ stable_baselines/__init__.py
@@ -7,7 +7,6 @@ from stable_baselines.ppo2 import PPO2
 from stable_baselines.td3 import TD3
 from stable_baselines.sac import SAC
 
-
 # Load mpi4py-dependent algorithms only if mpi is installed.
 try:
     import mpi4py
@@ -21,4 +20,4 @@ if mpi4py is not None:
     from stable_baselines.trpo_mpi import TRPO
 del mpi4py
 
-__version__ = "2.7.0"
+__version__ = "2.7.1a0"