+++ stable_baselines/common/vec_env/subproc_vec_env.py
@@ -55,17 +55,17 @@ class SubprocVecEnv(VecEnv):
     .. warning::
 
         Only 'forkserver' and 'spawn' start methods are thread-safe,
-        which is important when TensorFlow
-        sessions or other non thread-safe libraries are used in the parent (see issue #217).
-        However, compared to 'fork' they incur a small start-up cost and have restrictions on
-        global variables. With those methods,
-        users must wrap the code in an ``if __name__ == "__main__":``
+        which is important when TensorFlow sessions or other non thread-safe
+        libraries are used in the parent (see issue #217). However, compared to
+        'fork' they incur a small start-up cost and have restrictions on
+        global variables. With those methods, users must wrap the code in an
+        ``if __name__ == "__main__":`` block.
         For more information, see the multiprocessing documentation.
 
     :param env_fns: ([Gym Environment]) Environments to run in subprocesses
     :param start_method: (str) method used to start the subprocesses.
            Must be one of the methods returned by multiprocessing.get_all_start_methods().
-           Defaults to 'fork' on available platforms, and 'spawn' otherwise.
+           Defaults to 'forkserver' on available platforms, and 'spawn' otherwise.
     """
 
     def __init__(self, env_fns, start_method=None):
@@ -77,8 +77,8 @@ class SubprocVecEnv(VecEnv):
             # Fork is not a thread safe method (see issue #217)
             # but is more user friendly (does not require to wrap the code in
             # a `if __name__ == "__main__":`)
-            fork_available = 'fork' in multiprocessing.get_all_start_methods()
-            start_method = 'fork' if fork_available else 'spawn'
+            forkserver_available = 'forkserver' in multiprocessing.get_all_start_methods()
+            start_method = 'forkserver' if forkserver_available else 'spawn'
         ctx = multiprocessing.get_context(start_method)
 
         self.remotes, self.work_remotes = zip(*[ctx.Pipe() for _ in range(n_envs)])